Creates dev-requirements.txt pip install file
Provides the testinfra, pyest, and ansible versions necessary to
provision the SecureDrop VMs in the development environment.
Version-locking an older version of Ansible intentionally, in order to
stay as close as possible to what Admins are currently using via the
Tails apt repos (based on Debian Jessie).

Also updates the `pip_update.sh` script to update the developer
dependencies, via the pip-compile strategy used for managing the other
requirements lists.