Creates webassets directory in securedrop-app-code package
The use of Flask-Assets via #1180 for minifying static assets requires the
static asset directories to be present in the docroot. Although the
AppArmor rules permit writing inside of these directories already (those
changes were correctly included in #1180), the directories must already
exist, since the `mkdir` logic will fail, since the parent directory is
not writable in the AppArmor profile.

So the deb package should include the empty directories, but since the
server itself will handle minification, we should not ship any
pre-minified static assets; so let's add a config test to ensure that
doesn't happen. The config tests check for the existence of the
directories, and also that the directories are empty.